pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "an agentic RAG chatbot"
authors = ["Rosario Laface <lafacerosario@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.7.1"
openai = "^1.30.1"
langchain = "^0.2.0"
python-dotenv = "^1.0.1"
telepot = "^12.7"
qdrant-client = "^1.9.1"
groq = "^0.6.0"
langchain-groq = "^0.1.4"
langgraph = "^0.0.53"
fastembed = "^0.2.7"
tavily-python = "^0.3.3"
langchain-community = "^0.2.0"
boto3 = "^1.34.109"
config = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
